+++ reddwarf/data_presenter.py
@@ -7,18 +7,30 @@ from reddwarf.types.polis import PolisRepness
 import numpy as np
 import seaborn as sns
 
-from reddwarf.implementations.polis import PolisClusteringResult
+from reddwarf.implementations.base import PolisClusteringResult
 from reddwarf.utils.consensus import ConsensusResult
 
 GROUP_LABEL_NAMES = ["A", "B", "C", "D", "E", "F", "G", "H"]
 
 
 def generate_figure_polis(
-    result: PolisClusteringResult, show_guesses=False, flip_x=True, flip_y=False
+    result: PolisClusteringResult,
+    show_guesses=False,
+    flip_x=True,
+    flip_y=False,
+    show_pids=True,
 ):
-    cluster_labels = result.projected_participants["cluster_id"].values
+    """
+    Args:
+        show_pids (bool): Show the participant IDs on the plot
+    """
+    participants_clustered_df = result.participants_df[
+        result.participants_df["cluster_id"].notnull()
+    ]
+    cluster_labels = participants_clustered_df["cluster_id"].values
 
-    coord_data = result.projected_participants.loc[:, ["x", "y"]].values
+    coord_data = participants_clustered_df.loc[:, ["x", "y"]].values
+    coord_labels = None
     # Add the init center guesses to the bottom of the coord stack. Internally, they
     # will be give a fake "-1" colored label that won't be used to draw clusters.
     # This is for illustration purpose to see the centroid guesses.
@@ -26,13 +38,18 @@ def generate_figure_polis(
         coord_data = np.vstack(
             [
                 coord_data,
-                np.asarray(result.kmeans.init_centers_used_ if result.kmeans else []),
+                np.asarray(
+                    result.clusterer.init_centers_used_ if result.clusterer else []
+                ),
             ]
         )
 
+    if show_pids:
+        coord_labels = [f"p{pid}" for pid in participants_clustered_df.index]
+
     generate_figure(
         coord_data=coord_data,
-        coord_labels=[f"p{pid}" for pid in result.projected_participants.index],
+        coord_labels=coord_labels,
         cluster_labels=cluster_labels,
         # Always needs flipping to look like Polis interface.
         flip_x=flip_x,
@@ -99,11 +116,11 @@ def generate_figure(
         scatter_kwargs["cmap"] = "Set1"  # color map
 
         # Pad cluster_labels to match the number of points
-        UNGROUPED_LABEL = -1
+        CLUSTER_CENTER_LABEL = -2
         if len(cluster_labels) < len(coord_data):
             pad_length = len(coord_data) - len(cluster_labels)
             cluster_labels = np.concatenate(
-                [cluster_labels, [UNGROUPED_LABEL] * pad_length]
+                [cluster_labels, [CLUSTER_CENTER_LABEL] * pad_length]
             )
 
         scatter_kwargs["c"] = cluster_labels  # color indexes
@@ -112,8 +129,8 @@ def generate_figure(
         # Subset to allow unlabelled points to just be plotted
         unique_labels = np.unique(cluster_labels)
         for label in unique_labels:
-            if label == UNGROUPED_LABEL:
-                continue  # skip hulls when ungrouped label was padded in
+            if label in (-1, -2):
+                continue  # skip hulls when special-case labels used
 
             label_mask = cluster_labels == label
             cluster_points = coord_data[label_mask]
@@ -141,13 +158,17 @@ def generate_figure(
 
     # Add a legend if labels are provided
     if cluster_labels is not None:
-        cbar = plt.colorbar(scatter, label="Cluster", ticks=cluster_labels)
+        unique_labels = np.unique(cluster_labels)
+        cbar = plt.colorbar(scatter, label="Cluster", ticks=unique_labels)
 
-        UNGROUPED_LABEL_NAME = "[Center Guess]"
-        tick_labels = [
-            UNGROUPED_LABEL_NAME if lbl == -1 else GROUP_LABEL_NAMES[lbl]
-            for lbl in cluster_labels
-        ]
+        tick_labels = []
+        for lbl in unique_labels:
+            if lbl == -1:
+                tick_labels.append("[Unclustered]")
+            elif lbl == -2:
+                tick_labels.append("[Center Guess]")
+            else:
+                tick_labels.append(GROUP_LABEL_NAMES[lbl])
         cbar.ax.set_yticklabels(tick_labels)
 
     plt.show()
@@ -264,35 +285,11 @@ def print_repness(
         print("")
 
 
-class DataPresenter:
-    def __init__(self, client):
-        self.client = client
-
-    def render_optimal_cluster_figure(self):
-        print(f"Optimal clusters for K={self.client.optimal_k}")
-        print(
-            "Plotting PCA embeddings with K-means, K="
-            + str(np.max(self.client.optimal_cluster_labels) + 1)
-        )
-        self.generate_figure(
-            coord_dataframe=self.client.projected_data,
-            cluster_labels=self.client.optimal_cluster_labels,
-        )
-
-    def generate_figure(self, coord_dataframe, cluster_labels=None):
-        coord_data = coord_dataframe.loc[:, ["x", "y"]].values
-        coord_labels = coord_dataframe.index
-        generate_figure(
-            coord_data=coord_data,
-            coord_labels=coord_labels,
-            cluster_labels=cluster_labels,
-        )
-
-    def generate_vote_heatmap(self, vote_df):
-        sns.set_context("poster")
-        sns.set_style("white")
-        sns.set_theme(font_scale=0.7)
-        sns.set_color_codes()
-        fig, ax = plt.subplots(figsize=(10, 8))
-        sns.heatmap(vote_df, center=0, cmap="RdYlBu", ax=ax)
-        plt.show()
+def generate_vote_heatmap(vote_df):
+    sns.set_context("poster")
+    sns.set_style("white")
+    sns.set_theme(font_scale=0.7)
+    sns.set_color_codes()
+    fig, ax = plt.subplots(figsize=(10, 8))
+    sns.heatmap(vote_df, center=0, cmap="RdYlBu", ax=ax)
+    plt.show()