+++ reddwarf/utils/stats.py
@@ -9,7 +9,7 @@ from reddwarf.types.polis import (
     PolisRepness,
     PolisRepnessStatement,
 )
-from reddwarf.utils.pca import calculate_extremity
+from reddwarf.utils.reducer.pca import calculate_extremity
 
 
 def one_prop_test(
@@ -372,7 +372,7 @@ def format_comment_stats(statement: pd.Series) -> PolisRepnessStatement:
 
 def calculate_comment_statistics_dataframes(
     vote_matrix: VoteMatrix,
-    cluster_labels: list[int] | NDArray[np.integer],
+    cluster_labels: Optional[list[int] | NDArray[np.integer]] = None,
     pseudo_count: int = 1,
 ) -> Tuple[pd.DataFrame, pd.DataFrame]:
     """
@@ -397,6 +397,11 @@ def calculate_comment_statistics_dataframes(
             pseudo_count=pseudo_count,
         )
     )
+
+    if cluster_labels is None:
+        # Make a single group if no labels supplied.
+        participant_count = len(vote_matrix.index)
+        cluster_labels = [0] * participant_count
 
     group_count = len(set(cluster_labels))
     group_frames = []
@@ -438,7 +443,11 @@ def calculate_comment_statistics_dataframes(
     )
 
     group_aware_consensus_df = pd.DataFrame(
-        {"consensus": C_v_c[votes.A, :]},
+        {
+            "group-aware-consensus": C_v_c[votes.A, :],
+            "group-aware-consensus-agree": C_v_c[votes.A, :],
+            "group-aware-consensus-disagree": C_v_c[votes.D, :]
+        },
         index=vote_matrix.columns,
     )